refactor: combine delete logic to single hook
